requirements.txt
Django==3.2.5
djangorestframework==3.12.4
drf-yasg==1.20.0
Pillow==8.2.0